Give this some drip.
Sheesh.

+++ Client-2.0.0-github/Client-2.0.0/src/main/java/client/mixin/MixinPlayerControllerMP.java
@@ -58,7 +58,7 @@ public abstract class MixinPlayerControllerMP {
     @Inject(method={"getBlockReachDistance"}, at={@At(value="RETURN")}, cancellable=true)
     private void getReachDistanceHook(CallbackInfoReturnable<Float> distance) {
         if (Interactions.getInstance().isOn() && Interactions.getInstance().reach.getValue()) {
-            float range = distance.getReturnValue().floatValue();
+            float range = distance.getReturnValue ( );
             distance.setReturnValue(Interactions.getInstance().reachAmount.getValue());
         }
     }
